Added GraphBuilderTemplate
