+++ solvers/root_node/node.py
@@ -79,3 +79,4 @@ class RootNode:
 
     def color_reversed(self, x):
         return (x[2], x[1], x[0])
+
